daemon_kafka_consumer/requirements.txt
elasticsearch==8.3.1
kafka-python==2.0.0
requests==2.28.2
